+++ src/navbar/index.ts
@@ -1,2 +1,2 @@
-export { NavbarV2 } from "./navbar";
+export { Navbar } from "./navbar";
 export { BaseNavbar } from "./components/base-navbar";