+++ src/navbar/navbar.tsx
@@ -35,7 +35,7 @@ type Props = {
   currentPath: string;
 };
 
-export const NavbarV2: React.FC<Props> = props => {
+export const Navbar: React.FC<Props> = props => {
   const [isSidebarOpen, setIsSidebarOpen] = useState(false);
 
   const toggleSidebar = (): void => {