+++ src/table/index.ts
@@ -1,4 +1,3 @@
 export { PagedTable } from "./paged-table";
 export { SimpleTable } from "./simple-table";
 export { LargeSimpleTable } from "./large-simple-table";
-export { ExportTableButton } from "./components/export-table-button";